Move preview helpers to one place
